pyproject.toml
[project]
name = "newsvine-api"
version = "0.1.0"
description = "Local-first API for the Newsvine personal project"
readme = "LOCAL_SETUP.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "bcrypt>=4.2.0",
  "PyJWT>=2.9.0",
  "pydantic-settings>=2.3.4",
  "email-validator>=2.2.0",
  "requests>=2.32.3",
  "redis>=5.1.0",
  "kafka-python>=2.0.2",
  "feedparser>=6.0.11",
  "scikit-learn>=1.5.1",
  "mlflow>=2.16.2",
  "structlog>=24.4.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "opentelemetry-api>=1.28.0",
  "opentelemetry-sdk>=1.28.0",
  "opentelemetry-exporter-otlp>=1.28.0",
  "opentelemetry-instrumentation-fastapi>=0.49b0",
  "opentelemetry-instrumentation-requests>=0.49b0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "httpx>=0.27.0",
  "ruff>=0.5.7",
  "mypy>=1.11.1",
  "types-PyYAML>=6.0.12.20240311",
  "types-requests>=2.32.0.20240712"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
  "integration: marks tests that require running local services"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
namespace_packages = true
mypy_path = ["src"]
